refactor ts folder
